// packages/plugin-zenrock/src/types/zenrock/treasury/zrchain/key.ts
export enum KeyType {
  /** KEY_TYPE_UNSPECIFIED - The key type is missing. */
  KEY_TYPE_UNSPECIFIED = 0,
  /** KEY_TYPE_ECDSA_SECP256K1 - The key is an ECDSA secp256k1 key. */
  KEY_TYPE_ECDSA_SECP256K1 = 1,
  /** KEY_TYPE_EDDSA_ED25519 - The key is an EdDSA Ed25519 key. */
  KEY_TYPE_EDDSA_ED25519 = 2,
  /** KEY_TYPE_BITCOIN_SECP256K1 - Standard Bitcoin Keys */
  KEY_TYPE_BITCOIN_SECP256K1 = 3,
  UNRECOGNIZED = -1,
}